Reorganizing files and cleaning up globals from shadcn
