+++ src/app/components/ui/RecommendationCard.tsx
