+++ src/app/page.tsx
@@ -1,6 +1,8 @@
+"use client";
+
 import { useState, useMemo } from "react";
-import RecommendationCard from "../components/ui/RecommendationCard";
-import RecordResult from "../components/ui/RecordResult";
+import RecommendationCard from "./components/ui/RecommendationCard";
+import RecordResult from "./components/ui/RecordResult";
 import { recordResultsList } from "./utils/RecordResultsList";
 
 export default function Home() {
